+++ cmd/wharf/root.go
@@ -75,7 +75,7 @@ func loadKubeconfig() (*rest.Config, string, error) {
 func execute(version app.Version) {
 	var err error
 	if rootConfig, err = config.LoadConfig(); err != nil {
-		log.Error().Message(fmt.Sprintf("Config load: %s", err.Error()))
+		log.Error().Messagef("Config load: %s", err)
 		os.Exit(exitCodeLoadConfigError)
 	}