+++ pkg/aggregator/k8saggregator.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"net/http"
 	"net/url"
+	"strconv"
 	"sync"
 	"time"
 
@@ -13,7 +14,7 @@ import (
 	"github.com/iver-wharf/wharf-cmd/pkg/config"
 	"github.com/iver-wharf/wharf-cmd/pkg/workerapi/workerclient"
 	"github.com/iver-wharf/wharf-core/pkg/logger"
-	"gopkg.in/typ.v3/pkg/sets"
+	"gopkg.in/typ.v4/maps"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 	"k8s.io/client-go/tools/portforward"
@@ -98,7 +99,7 @@ func (a k8sAggr) Serve(ctx context.Context) error {
 	k8sruntime.ErrorHandlers = []func(error){}
 
 	var mutex sync.Mutex
-	inProgress := make(sets.Set[types.UID])
+	inProgress := make(maps.Set[types.UID])
 	for {
 		// TODO: Wait for Wharf API to be up first, with sane infinite retry logic.
 		//
@@ -121,6 +122,13 @@ func (a k8sAggr) Serve(ctx context.Context) error {
 			if pod.Status.Phase != v1.PodRunning {
 				continue
 			}
+			buildID, err := parsePodBuildID(pod.ObjectMeta)
+			if err != nil {
+				log.Warn().WithError(err).
+					WithStringf("pod", "%s/%s", pod.Namespace, pod.Name).
+					Message("Failed to parse worker's build ID.")
+				continue
+			}
 			if inProgress.Has(pod.UID) {
 				continue
 			}
@@ -129,18 +137,32 @@ func (a k8sAggr) Serve(ctx context.Context) error {
 				WithStringf("pod", "%s/%s", pod.Namespace, pod.Name).
 				Message("Pod found.")
 
-			go func(p v1.Pod) {
-				if err := a.relayToWharfAPI(ctx, p.Name); err != nil {
-					log.Error().WithError(err).Message("Relay error.")
+			go func(pod v1.Pod) {
+				if err := a.relayToWharfAPI(ctx, pod.Name, buildID); err != nil {
+					log.Error().WithError(err).
+						WithStringf("pod", "%s/%s", pod.Namespace, pod.Name).
+						Message("Relay error.")
 				}
 				mutex.Lock()
-				inProgress.Remove(p.UID)
+				inProgress.Remove(pod.UID)
 				mutex.Unlock()
 			}(pod)
 		}
 		mutex.Unlock()
 		time.Sleep(pollDelay)
 	}
+}
+
+func parsePodBuildID(podMeta metav1.ObjectMeta) (uint, error) {
+	buildRef, ok := podMeta.Labels["wharf.iver.com/build-ref"]
+	if !ok {
+		return 0, errors.New("missing label 'wharf.iver.com/build-ref'")
+	}
+	buildID, err := strconv.ParseUint(buildRef, 10, 0)
+	if err != nil {
+		return 0, err
+	}
+	return uint(buildID), nil
 }
 
 func (a k8sAggr) fetchPods(ctx context.Context) ([]v1.Pod, error) {
@@ -163,14 +185,14 @@ func (a k8sAggr) fetchPods(ctx context.Context) ([]v1.Pod, error) {
 	return pods, nil
 }
 
-func (a k8sAggr) relayToWharfAPI(ctx context.Context, podName string) error {
+func (a k8sAggr) relayToWharfAPI(ctx context.Context, podName string, buildID uint) error {
 	portConn, err := a.newPortForwarding(a.namespace, podName)
 	if err != nil {
 		return err
 	}
 	defer portConn.Close()
 
-	worker, err := a.newWorkerClient(portConn)
+	worker, err := a.newWorkerClient(portConn, buildID)
 	if err != nil {
 		return err
 	}
@@ -188,9 +210,6 @@ func (a k8sAggr) relayToWharfAPI(ctx context.Context, podName string) error {
 		return fmt.Errorf("relaying to wharf: %w", err)
 	}
 
-	// TODO: Check build results from already-streamed status events if the
-	// build is actually done. If not, then set build as failed via wharfapi.
-
 	log.Debug().
 		WithStringf("pod", "%s/%s", a.namespace, podName).
 		Message("Done relaying. Terminating pod.")
@@ -205,13 +224,14 @@ func (a k8sAggr) relayToWharfAPI(ctx context.Context, podName string) error {
 	return nil
 }
 
-func (a k8sAggr) newWorkerClient(portConn portConnection) (workerclient.Client, error) {
+func (a k8sAggr) newWorkerClient(portConn portConnection, buildID uint) (workerclient.Client, error) {
 	// Intentionally "localhost" because we're port-forwarding
 	return workerclient.New(fmt.Sprintf("http://localhost:%d", portConn.Local), workerclient.Options{
 		// Skipping security because we've already authenticated with Kubernetes
 		// and are communicating through a secured port-forwarding tunnel.
 		// Don't need to add TLS on top of TLS.
 		InsecureSkipVerify: true,
+		BuildID:            buildID,
 	})
 }